requirements.txt
telethon
ProfanityDetector
python-decouple
requests